+++ XercaFood/src/main/java/xerca/xercafood/common/block_entity/BlockEntityDoner.java
@@ -1,5 +1,6 @@
 package xerca.xercafood.common.block_entity;
 
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.sounds.SoundSource;
@@ -8,7 +9,6 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import xerca.xercafood.common.SoundEvents;
 
 import static net.minecraft.world.level.block.CampfireBlock.LIT;
@@ -44,7 +44,7 @@ public class BlockEntityDoner extends BlockEntity {
     }
 
     public float getAnimationProgress(float partialTicks) {
-        return this.isSpinning ? (float) this.spinTicks + partialTicks : (float) this.spinTicks;
+        return this.isSpinning ? this.spinTicks + partialTicks : this.spinTicks;
     }
 
     @SuppressFBWarnings(value = "NP", justification = "gettingRoasted is used during block entity ticking when level is attached.")
