+++ XercaFood/src/main/java/xerca/xercafood/common/item/ItemEnderCupcake.java
@@ -28,7 +28,7 @@ class ItemEnderCupcake extends Item {
 
             for (int lvt_11_1_ = 0; lvt_11_1_ < 16; ++lvt_11_1_) {
                 double x = entity.getX() + (entity.getRandom().nextDouble() - 0.5D) * 8.0D;
-                double y = Mth.clamp(entity.getY() + (double) (entity.getRandom().nextInt(8) - 4), 0.0D, (world.getHeight() - 1));
+                double y = Mth.clamp(entity.getY() + (entity.getRandom().nextInt(8) - 4), 0.0D, (world.getHeight() - 1));
                 double z = entity.getZ() + (entity.getRandom().nextDouble() - 0.5D) * 8.0D;
                 if (entity.isPassenger()) {
                     entity.stopRiding();
